Fix implicit type conversion.

+++ src/random.cpp
@@ -8,18 +8,18 @@
 
 #include "random.h"
 
-int Random::id = 0;
+uint_fast64_t Random::id = 0;
 
 Random::Random(double mean, double std) {
   init(++id,mean,std);
 }
 
-Random::Random(int seed, double mean, double std) {
+Random::Random(uint_fast64_t seed, double mean, double std) {
   init(seed,mean,std);
   ++id;
 }
 
-void Random::init(int seed, double mean, double std) {
+void Random::init(uint_fast64_t seed, double mean, double std) {
   gen = std::mt19937_64(seed); // Generator
   dist = std::normal_distribution<double>(mean,std); // Distribution
 }
